Do the "-local1" version suffix in a better way
Instead of editing __version__ in lib/sqlalchemy/__init__.py,
we leave that at its upstream value ("0.9.8"), and set
"tag_build = -local1" in setup.cfg.

It was the upstream value of that, "tag_build = dev", which was
responsible for making sdists like "SQLAlchemy-0.9.8-local1dev" (note
the "dev" suffix), which in turn meant that pip couldn't find the
SQLAlchemy-0.9.8-local1 (note: no "dev" suffix) package *as a dependency
of cma*, which in turn broke tox testing.

Judging by tag_build's upstream value, "dev", I think that it should
ideally be set to "local1", not "-local1".  But the latter is needed for
backward compatibility with my previous way of doing it, i.e.  so that
old versions of CMA can continue to find this version of SQLAlchemy, for
bug bisection.